# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Supply/demand zone intelligence API. Provides zone detection, live price streaming, market metrics, and MCP tool access for AI agents.

 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export type ZoneDirection = (typeof ZoneDirection)[keyof typeof ZoneDirection];

export const ZoneDirection = {
  supply: "supply",
  demand: "demand",
} as const;

export interface Zone {
  id: number;
  symbol: string;
  direction: ZoneDirection;
  timeframe: string;
  proximalLine: number;
  distalLine: number;
  combinedConfidence: number;
  priceInside: boolean;
  active: boolean;
}

export type ConfluentZoneDirection =
  (typeof ConfluentZoneDirection)[keyof typeof ConfluentZoneDirection];

export const ConfluentZoneDirection = {
  supply: "supply",
  demand: "demand",
} as const;

export interface ConfluentZone {
  id: number;
  symbol: string;
  direction: ConfluentZoneDirection;
  timeframes: string[];
  proximalLine: number;
  distalLine: number;
  combinedConfidence: number;
  priceInside: boolean;
}

export interface RefreshResult {
  ok: boolean;
  count: number;
}

export interface MonitoredSymbol {
  symbol: string;
  currentPrice?: number | null;
  zoneCount: number;
}

export interface SymbolStatus {
  symbol: string;
  status: string;
}

export interface PriceResult {
  symbol: string;
  price: number | null;
}

export interface Candle {
  timestamp: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface ScanResult {
  symbol: string;
  ivRank?: number;
  ivx?: number;
  earningsDate?: string | null;
  sectorIvRank?: number | null;
}

/**
 * Unauthorized — missing or invalid INTERNAL_API_TOKEN
 */
export type UnauthorizedResponse = void;

export type GetTopZonesParams = {
  /**
   * Max zones to return (default 10).
   */
  limit?: number;
  /**
   * Filter by zone direction.
   */
  direction?: GetTopZonesDirection;
};

export type GetTopZonesDirection =
  (typeof GetTopZonesDirection)[keyof typeof GetTopZonesDirection];

export const GetTopZonesDirection = {
  supply: "supply",
  demand: "demand",
} as const;

export type GetZonesBySymbolParams = {
  /**
   * Filter by timeframe, e.g. 60m, 1d.
   */
  timeframe?: string;
};