# lib/api-zod/src/generated/types/refreshResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Supply/demand zone intelligence API. Provides zone detection, live price streaming, market metrics, and MCP tool access for AI agents.

 * OpenAPI spec version: 1.0.0
 */

export interface RefreshResult {
  ok: boolean;
  count: number;
}